KSieve Session: fix race while handling SSL errors.
Summary:
No need to post a request to the thread for handling an ignored ssl error,
which the thread will simply handle by emitting the sslDone signal,
which goes back to the main thread. By doing that we were making it possible
for the thread to emit other things meanwhile - like the received data from the server,
out of sequence (the main thread can only handle it after receiving sslDone)

This fixes the never-ending spinning in "Manage Sieve Scripts" in kmail
for me (which happened about 20% of the time).

BUG: 232246

Test Plan: Manage Sieve Scripts in kmail, > 5 times.

Reviewers: dvratil, mlaurent

Reviewed By: mlaurent

Subscribers: kde-pim, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D2167